Handling errors with Promise.allSettled
Wanting as much data as possible to come back.
If one or maybe even two promises fail, that's okay so long as we get the rest of the data.

OBS:

if Promise.allSettled if 'fulfilled' is true, we get 'value'
if 'refejected' is true, we get reason (e.g. reason.message)